Delete file database_connect.py
