+++ results/DIETClassifier_report.json
@@ -1,27 +1,27 @@
 {
   "c++_content": {
-    "precision": 1.0,
-    "recall": 0.9813829787234043,
-    "f1-score": 0.9906040268456376,
-    "support": 376,
+    "precision": 0.9181969949916527,
+    "recall": 0.8785942492012779,
+    "f1-score": 0.8979591836734695,
+    "support": 1252,
     "confused_with": {}
   },
   "micro avg": {
-    "precision": 1.0,
-    "recall": 0.9813829787234043,
-    "f1-score": 0.9906040268456376,
-    "support": 376
+    "precision": 0.9181969949916527,
+    "recall": 0.8785942492012779,
+    "f1-score": 0.8979591836734695,
+    "support": 1252
   },
   "macro avg": {
-    "precision": 1.0,
-    "recall": 0.9813829787234043,
-    "f1-score": 0.9906040268456376,
-    "support": 376
+    "precision": 0.9181969949916527,
+    "recall": 0.8785942492012779,
+    "f1-score": 0.8979591836734695,
+    "support": 1252
   },
   "weighted avg": {
-    "precision": 1.0,
-    "recall": 0.9813829787234043,
-    "f1-score": 0.9906040268456376,
-    "support": 376
+    "precision": 0.9181969949916529,
+    "recall": 0.8785942492012779,
+    "f1-score": 0.8979591836734695,
+    "support": 1252
   }
 }
