+++ results/intent_report.json
@@ -1,108 +1,13 @@
 {
-  "goodbye": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 18,
-    "confused_with": {}
-  },
-  "greet": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 39,
-    "confused_with": {}
-  },
-  "c++_why_asking": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 18,
-    "confused_with": {}
-  },
-  "deny": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 20,
-    "confused_with": {}
-  },
-  "compliment": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 28,
-    "confused_with": {}
-  },
-  "out_of_scope": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 69,
-    "confused_with": {}
-  },
-  "insult": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 74,
-    "confused_with": {}
-  },
-  "c++_what_asking": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 167,
-    "confused_with": {}
-  },
-  "c++_where_asking": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 14,
-    "confused_with": {}
-  },
-  "stop_asking": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 70,
-    "confused_with": {}
-  },
   "thanks": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
+    "precision": 0.9090909090909091,
+    "recall": 0.8333333333333334,
+    "f1-score": 0.8695652173913043,
     "support": 24,
-    "confused_with": {}
-  },
-  "health_inquire": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 83,
-    "confused_with": {}
-  },
-  "want_to_ask": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 31,
-    "confused_with": {}
-  },
-  "c++_how_asking": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 85,
-    "confused_with": {}
-  },
-  "affirm": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 17,
-    "confused_with": {}
+    "confused_with": {
+      "deny": 2,
+      "insult": 1
+    }
   },
   "c++_when_asking": {
     "precision": 1.0,
@@ -111,38 +16,228 @@
     "support": 21,
     "confused_with": {}
   },
-  "feeling_bad": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 45,
-    "confused_with": {}
+  "health_inquire": {
+    "precision": 0.9097744360902256,
+    "recall": 0.9603174603174603,
+    "f1-score": 0.9343629343629345,
+    "support": 126,
+    "confused_with": {
+      "feeling_good": 2,
+      "who_are_you": 1
+    }
   },
   "continue_asking": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
+    "precision": 0.875,
+    "recall": 0.9210526315789473,
+    "f1-score": 0.8974358974358975,
     "support": 38,
+    "confused_with": {
+      "affirm": 1,
+      "stop_asking": 1
+    }
+  },
+  "attention": {
+    "precision": 0.4146341463414634,
+    "recall": 0.425,
+    "f1-score": 0.41975308641975306,
+    "support": 40,
+    "confused_with": {
+      "greet": 14,
+      "insult": 4
+    }
+  },
+  "insult": {
+    "precision": 0.8846153846153846,
+    "recall": 0.8761904761904762,
+    "f1-score": 0.8803827751196172,
+    "support": 105,
+    "confused_with": {
+      "attention": 2,
+      "c++_what_asking": 2
+    }
+  },
+  "dont_want_to_ask": {
+    "precision": 0.8166666666666667,
+    "recall": 0.875,
+    "f1-score": 0.8448275862068966,
+    "support": 56,
+    "confused_with": {
+      "c++_what_asking": 3,
+      "deny": 1
+    }
+  },
+  "feeling_bad": {
+    "precision": 0.8809523809523809,
+    "recall": 0.8222222222222222,
+    "f1-score": 0.8505747126436781,
+    "support": 45,
+    "confused_with": {
+      "deny": 2,
+      "affirm": 2
+    }
+  },
+  "who_are_you": {
+    "precision": 0.9174311926605505,
+    "recall": 0.9803921568627451,
+    "f1-score": 0.9478672985781992,
+    "support": 102,
+    "confused_with": {
+      "c++_what_asking": 1,
+      "greet": 1
+    }
+  },
+  "greet": {
+    "precision": 0.6065573770491803,
+    "recall": 0.6851851851851852,
+    "f1-score": 0.6434782608695653,
+    "support": 54,
+    "confused_with": {
+      "attention": 14,
+      "compliment": 2
+    }
+  },
+  "want_to_ask": {
+    "precision": 0.90625,
+    "recall": 0.9354838709677419,
+    "f1-score": 0.9206349206349206,
+    "support": 31,
+    "confused_with": {
+      "c++_what_asking": 2
+    }
+  },
+  "corona": {
+    "precision": 0.9761904761904762,
+    "recall": 1.0,
+    "f1-score": 0.9879518072289156,
+    "support": 41,
     "confused_with": {}
   },
   "feeling_good": {
-    "precision": 1.0,
+    "precision": 0.8,
+    "recall": 0.6666666666666666,
+    "f1-score": 0.7272727272727272,
+    "support": 54,
+    "confused_with": {
+      "health_inquire": 7,
+      "compliment": 4
+    }
+  },
+  "c++_why_asking": {
+    "precision": 0.9631336405529954,
+    "recall": 0.9812206572769953,
+    "f1-score": 0.972093023255814,
+    "support": 213,
+    "confused_with": {
+      "c++_what_asking": 2,
+      "c++_how_asking": 1
+    }
+  },
+  "deny": {
+    "precision": 0.7611940298507462,
+    "recall": 0.7611940298507462,
+    "f1-score": 0.7611940298507462,
+    "support": 67,
+    "confused_with": {
+      "dont_want_to_ask": 4,
+      "stop_asking": 3
+    }
+  },
+  "goodbye": {
+    "precision": 0.9285714285714286,
+    "recall": 0.5909090909090909,
+    "f1-score": 0.7222222222222223,
+    "support": 22,
+    "confused_with": {
+      "deny": 2,
+      "continue_asking": 2
+    }
+  },
+  "compliment": {
+    "precision": 0.36666666666666664,
+    "recall": 0.39285714285714285,
+    "f1-score": 0.3793103448275862,
+    "support": 28,
+    "confused_with": {
+      "feeling_good": 5,
+      "affirm": 2
+    }
+  },
+  "stop_asking": {
+    "precision": 0.8831168831168831,
+    "recall": 0.9714285714285714,
+    "f1-score": 0.9251700680272108,
+    "support": 70,
+    "confused_with": {
+      "insult": 1,
+      "dont_want_to_ask": 1
+    }
+  },
+  "c++_where_asking": {
+    "precision": 0.9333333333333333,
     "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 40,
+    "f1-score": 0.9655172413793104,
+    "support": 14,
     "confused_with": {}
   },
-  "accuracy": 1.0,
-  "macro avg": {
+  "out_of_scope": {
+    "precision": 0.9375,
+    "recall": 0.6521739130434783,
+    "f1-score": 0.7692307692307693,
+    "support": 69,
+    "confused_with": {
+      "who_are_you": 3,
+      "dont_want_to_ask": 3
+    }
+  },
+  "c++_what_asking": {
+    "precision": 0.9525,
+    "recall": 0.9719387755102041,
+    "f1-score": 0.9621212121212122,
+    "support": 392,
+    "confused_with": {
+      "c++_why_asking": 4,
+      "who_are_you": 3
+    }
+  },
+  "affirm": {
+    "precision": 0.625,
+    "recall": 0.4166666666666667,
+    "f1-score": 0.5,
+    "support": 24,
+    "confused_with": {
+      "compliment": 3,
+      "continue_asking": 2
+    }
+  },
+  "cpp_practice": {
     "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 901
+    "recall": 0.9545454545454546,
+    "f1-score": 0.9767441860465117,
+    "support": 22,
+    "confused_with": {
+      "dont_want_to_ask": 1
+    }
+  },
+  "c++_how_asking": {
+    "precision": 0.9782608695652174,
+    "recall": 0.989010989010989,
+    "f1-score": 0.9836065573770493,
+    "support": 91,
+    "confused_with": {
+      "c++_what_asking": 1
+    }
+  },
+  "accuracy": 0.8850771869639794,
+  "macro avg": {
+    "precision": 0.8427683258881046,
+    "recall": 0.8192828872676715,
+    "f1-score": 0.826721536604285,
+    "support": 1749
   },
   "weighted avg": {
-    "precision": 1.0,
-    "recall": 1.0,
-    "f1-score": 1.0,
-    "support": 901
+    "precision": 0.8856560109572129,
+    "recall": 0.8850771869639794,
+    "f1-score": 0.8829826477546328,
+    "support": 1749
   }
 }
